requirements.txt
django==5.2.4
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
python-decouple==3.8
Pillow==11.0.0